cleanup and no (functioless) buttons
