+++ src/point_tree_item.cpp
@@ -1,6 +1,6 @@
 #include <QStringList>
 
- #include <moveit_cartesian_plan_plugin/point_tree_item.hpp>
+ #include <moveit_cartesian_plan_plugin/point_tree_item.h>
 
 PointTreeItem::PointTreeItem(const QVector<QVariant> &data, PointTreeItem *parent)
 {
