+++ src/widgets/path_planning_widget.ui
@@ -7,7 +7,7 @@
     <x>0</x>
     <y>0</y>
     <width>428</width>
-    <height>742</height>
+    <height>933</height>
    </rect>
   </property>
   <property name="windowTitle">
@@ -28,28 +28,65 @@
       </attribute>
       <layout class="QVBoxLayout" name="verticalLayout_7">
        <item>
-        <layout class="QVBoxLayout" name="verticalLayout_2">
+        <widget class="QLabel" name="label_32">
+         <property name="font">
+          <font>
+           <pointsize>11</pointsize>
+           <weight>75</weight>
+           <bold>true</bold>
+          </font>
+         </property>
+         <property name="text">
+          <string>Planning Environment</string>
+         </property>
+        </widget>
+       </item>
+       <item>
+        <layout class="QGridLayout" name="gridLayout_8" columnstretch="48,4,48">
          <property name="margin">
           <number>0</number>
          </property>
-         <item>
-          <widget class="QLabel" name="label_32">
-           <property name="font">
-            <font>
-             <pointsize>11</pointsize>
-             <weight>75</weight>
-             <bold>true</bold>
-            </font>
-           </property>
+         <item row="0" column="0">
+          <widget class="QLabel" name="planningGroupLabel">
            <property name="text">
-            <string>Planning Group</string>
+            <string>Move Group</string>
            </property>
           </widget>
          </item>
-         <item>
+         <item row="0" column="2">
           <widget class="QComboBox" name="combo_planGroup">
            <property name="toolTip">
             <string>Select a planning group for MoveIt!</string>
+           </property>
+          </widget>
+         </item>
+         <item row="0" column="1">
+          <spacer name="horizontalSpacer_2">
+           <property name="orientation">
+            <enum>Qt::Horizontal</enum>
+           </property>
+           <property name="sizeHint" stdset="0">
+            <size>
+             <width>40</width>
+             <height>20</height>
+            </size>
+           </property>
+          </spacer>
+         </item>
+         <item row="1" column="2">
+          <widget class="QLineEdit" name="referenceFrameInput">
+           <property name="enabled">
+            <bool>false</bool>
+           </property>
+           <property name="text">
+            <string>base_link</string>
+           </property>
+          </widget>
+         </item>
+         <item row="1" column="0">
+          <widget class="QLabel" name="referenceFrameLabel">
+           <property name="text">
+            <string>Reference Frame</string>
            </property>
           </widget>
          </item>
@@ -97,8 +134,18 @@
         </widget>
        </item>
        <item>
-        <layout class="QHBoxLayout" name="horizontalLayout_2" stretch="48,4,48">
-         <item>
+        <layout class="QGridLayout" name="gridLayout_7" columnstretch="48,4,48">
+         <item row="1" column="2">
+          <widget class="QPushButton" name="pushButton">
+           <property name="toolTip">
+            <string>Move selected waypoint up</string>
+           </property>
+           <property name="text">
+            <string>△</string>
+           </property>
+          </widget>
+         </item>
+         <item row="1" column="0">
           <widget class="QPushButton" name="moveToWaypointButton">
            <property name="toolTip">
             <string>Move arm to selected waypoint</string>
@@ -108,7 +155,27 @@
            </property>
           </widget>
          </item>
-         <item>
+         <item row="2" column="2">
+          <widget class="QPushButton" name="pushButton_2">
+           <property name="toolTip">
+            <string>Move selected waypoint down</string>
+           </property>
+           <property name="text">
+            <string>▽</string>
+           </property>
+          </widget>
+         </item>
+         <item row="2" column="0">
+          <widget class="QPushButton" name="deleteWaypointButton">
+           <property name="toolTip">
+            <string>Delete selected waypoint</string>
+           </property>
+           <property name="text">
+            <string>Delete Waypoint</string>
+           </property>
+          </widget>
+         </item>
+         <item row="1" column="1">
           <spacer name="horizontalSpacer">
            <property name="orientation">
             <enum>Qt::Horizontal</enum>
@@ -120,16 +187,6 @@
             </size>
            </property>
           </spacer>
-         </item>
-         <item>
-          <widget class="QPushButton" name="deleteWaypointButton">
-           <property name="toolTip">
-            <string>Delete selected waypoint</string>
-           </property>
-           <property name="text">
-            <string>Delete Waypoint</string>
-           </property>
-          </widget>
          </item>
         </layout>
        </item>
@@ -1205,6 +1262,9 @@
          </item>
          <item row="0" column="2">
           <widget class="QLineEdit" name="lnEdit_PlanTime">
+           <property name="text">
+            <string>5.0</string>
+           </property>
            <property name="alignment">
             <set>Qt::AlignRight|Qt::AlignTrailing|Qt::AlignVCenter</set>
            </property>
@@ -1269,6 +1329,9 @@
               <property name="styleSheet">
                <string notr="true"/>
               </property>
+              <property name="text">
+               <string>0.01</string>
+              </property>
               <property name="alignment">
                <set>Qt::AlignRight|Qt::AlignTrailing|Qt::AlignVCenter</set>
               </property>
@@ -1300,7 +1363,7 @@
                <string notr="true"/>
               </property>
               <property name="text">
-               <string/>
+               <string>0.0</string>
               </property>
               <property name="alignment">
                <set>Qt::AlignRight|Qt::AlignTrailing|Qt::AlignVCenter</set>
